requirements.txt
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
openpyxl==3.1.2
xlsxwriter==3.1.9
camelot-py[cv]==0.10.1
opencv-python==4.8.1.78
ghostscript==0.7